Sort registrations by status.
This should give us PreApproved registrations if multiple exist.
Also, as part of approving a preapproval, we also approve other preapproved requests.

There is still some scope for requests getting in the way of each other.
For this, we should start deleteing the requests manually and then slowly add the user
to their various facilities